feat: rename to VariableMissingErrorModal
